Working on valid expressions
Note that modulo SHOULD be on the same order as multiplication and division and be evaluated from left to right. However, multiplication and division don't really matter which order they are done in but modulo does. Right now, modulo is placed as a lower priority. A future update could add priority levels and then fallback on left to right evaluation for the same priority.

Next I want to add error checking