Incomplete Euclidean Algorithm Calculator.
Right now it has the basic algorithm implemented so it can calculate the greatest common factor between two integers. It also has a semi-functional function that finds the Bezout coefficients by stepping through the Euclidean Algorithm backwards. This is not done through the Extended Euclidean Algorithm.

It still needs to be formatted so the factors are written out. It also is not working with the inputs 414 and 662 for some reason. My best guess is that the subtraction is going wrong at some point as it takes 4 steps and the other test cases only take 2 or 4 steps.